Created common abstract activity that implements main option menu
Menu list, menu details and food details extended that activity
