+++ app/src/main/java/com/canteen/app/activity/client/food/details/FoodDetailsActivity.java
@@ -1,22 +1,15 @@
 package com.canteen.app.activity.client.food.details;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.widget.TextView;
 
 import com.canteen.app.R;
-import com.canteen.app.activity.client.cart.OrderCartActivity;
+import com.canteen.app.activity.client.common.ActivityWithMainOptionMenu;
 import com.canteen.app.models.Food;
 import com.canteen.app.service.ToastService;
-import com.canteen.app.service.login.LoginService;
-import com.canteen.app.service.login.LoginServiceImpl;
 import com.canteen.app.service.order.OrderCartService;
 import com.canteen.app.service.order.OrderItem;
 import com.canteen.app.service.price.PriceFormatter;
@@ -26,13 +19,11 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
-public class FoodDetailsActivity extends AppCompatActivity implements PriceContainer {
+public class FoodDetailsActivity extends ActivityWithMainOptionMenu implements PriceContainer {
 
     private Food food;
 
     private FoodAdditionAdapter adapter;
-
-    private LoginService loginService = LoginServiceImpl.of();
 
     @BindView(R.id.food_description_text_view)
     TextView foodDescriptionTextView;
@@ -52,28 +43,6 @@ public class FoodDetailsActivity extends AppCompatActivity implements PriceConta
         retrieveFood();
         initView();
         updatePrice(0);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(final Menu menu) {
-        super.onCreateOptionsMenu(menu);
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menu, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(final MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_cart:
-                startActivity(new Intent(this, OrderCartActivity.class));
-                return true;
-            case R.id.action_logout:
-                loginService.logout(this);
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
     }
 
     @OnClick(R.id.add_to_cart_button)
